==> Modelacion/1_MineriaDatos_Partición_de_datos.ipynb <==
## Notebook para realizar la partición de los datos

Este libro de jupyter se utiliza para realizar la partición de los datos que se van a emplear en el modelo, obteniendo de esta manera dos muestra, una de train que se utiliza para el entrenamiento y una de test con el fin de realizar la validación del modelo

# Importamos las librerias necesarias
import pandas as pd

# Realizamos la lectura de la información
x =pd.read_csv('10000.txt', 
               sep='\t', 
               header=None,
               names=['user','song','play_count'])

# Realizamos una rapida exploracion de la informacion cargada
x.describe

# Observamos los primeros registros del dataset
x.head()

# Creamos un dataframe para almacenar los datos de test
x_test = pd.DataFrame(columns=['user','song','play_count'])

# Se realiza una particion de los datos, tomando el 25% de la muestra
# original con el fin de utilizarlos para el testeo del modelo
%time

for i in x['user'].unique().tolist():
    x_1 = x[x["user"] == i].sample(frac=0.25)
    x_test = pd.concat([x_test, x_1])

# Almacenamos los datos de test en formato csv
x_test.to_csv('x_test.csv')

# Hacemos un renombramiento de una columna del df para mas adelante
# realizar un full outer join entre el dataset completo y el test
# para obtener los datos de train
x_testR=x_test.rename(columns={"play_count": "a"});

# Obtenemos los datos train
x_train=x.merge(x_testR, 
                on = ['user','song'] ,
                how='left',indicator = 'i').query('i == "left_only"').drop(['a', 'i'], axis=1)

# Y almacenamos finalmente los datos de train en formato csv
x_train.to_csv('x_train.csv')
